+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - 🎉 Added `id.nanoid`
 - 🎉 Added `color.human`
 - 🎉 Added `date.anytime`
+- 🎉 Added `finance.litecoinAddress`
+- 🎉 Added `internet.ip`
 - The `prefix` argument was added to the `datatype.alphaNumeric` module
 
 ## 🪛 Fix